Add python implementations
Add python implementations directory so that easy implementations can be
added to the repo.

This will allow more rapid implementations than translating the
pseudocode to a functional, immutable language like Elixir. While fun,
it's definitely slowing down progress and much time is spent on the
translation, rather than the learning of the algorithm/data structure.